requirements-dev.txt
coverage==5.4
cython==0.29.22
line_profiler==3.1.0
django-debug-toolbar==3.2
django-debug-toolbar-line-profiling==0.7.1
django-debug-toolbar-request-history==0.1.3
-e git://github.com/mikekeda/django-jenkins.git#egg=django-jenkins
pep8==1.7.1
pyflakes==2.2.0
pylint==2.7.1
pylint-django==2.4.2
pylint-plugin-utils==0.6
